Renamed fileGUI to MainView and created the FileView Class
